Add pagination functionality to the emails entity. Update the main file, so that it runs the project on different ports, if any are used, so that it can always run, if it is in development mode

+++ index.js
@@ -18,19 +18,69 @@ addAliases({
 const fs = require("fs")
 const path = require("path")
 
-if (fs.existsSync(path.join(__dirname, "tmp"))) {
-  fs.rmSync(path.join(__dirname, "tmp"), { recursive: true, force: true })
-}
-
 const { server } = require("./app.js")
 
-const PORT = process.env.PORT || 8080
+const PORT = Number(process.env.PORT) || 8080
 const isProduction = process.env.NODE_ENV === "production"
+const initialPort = PORT
 
-server.listen(PORT, () => {
-  if (isProduction) {
-    console.log("Server is on...")
-  } else {
-    console.log(`Listening on port ${PORT} (http://localhost:${PORT})...`)
+if (!isProduction) {
+  const isPortInUse = (port) => {
+    return new Promise((resolve) => {
+      const server = require("net").createServer()
+      server.once("error", (err) => {
+        if (err.code === "EADDRINUSE") {
+          resolve(true)
+        } else {
+          resolve(false)
+        }
+      })
+      server.once("listening", () => {
+        server.close()
+        resolve(false)
+      })
+      server.listen(port)
+    })
   }
-})
+
+  const startServer = async (port) => {
+    while (await isPortInUse(port)) {
+      console.warn(`Port ${port} is already in use. Trying next port...`)
+      await new Promise((resolve) => setTimeout(resolve, 2000))
+      port++
+    }
+
+    try {
+      if (port === initialPort && fs.existsSync(path.join(__dirname, "tmp"))) {
+        fs.rmSync(path.join(__dirname, "tmp"), { recursive: true, force: true })
+      }
+    } catch (error) {}
+
+    return new Promise((resolve, reject) => {
+      server.listen(port)
+      server.on("listening", () => {
+        resolve(port)
+      })
+      server.on("error", (error) => {
+        server.close()
+        reject(error)
+      })
+    })
+  }
+
+  startServer(PORT)
+    .then((usedPort) => {
+      console.log(`Listening on port ${usedPort} (http://localhost:${usedPort})...`)
+    })
+    .catch((error) => {
+      throw new Error(error)
+    })
+} else {
+  if (fs.existsSync(path.join(__dirname, "tmp"))) {
+    fs.rmSync(path.join(__dirname, "tmp"), { recursive: true, force: true })
+  }
+
+  server.listen(PORT, () => {
+    console.log("Server is running...")
+  })
+}
